+++ client/src/container/CategoryManageContainer/styles.tsx
@@ -5,6 +5,9 @@ const CategoryListContainer = styled.div`
   flex-direction: column;
   width: 100%;
   margin-top: 1.5rem;
+  border-radius: 0.3rem;
+  box-shadow: 1px 2px 4px 1px rgba(0, 0, 0, 0.2);
+  padding: 1rem;
 `;
 
 export default CategoryListContainer;
